Refactor - image onError - itemsPage :  separation of responsibility
